refactor: mv (big) util funcs out of comp files
Why?
To reduce clutter.
