+++ src/utils/getInputValueFromQuery.ts
@@ -1,0 +1,6 @@
+import { searchInput } from "@/utils/viewInputsData";
+import type { Query } from "@/utils/types";
+
+export const getInputValueFromQuery = (query: Query) => {
+  return query[searchInput.name] ?? "";
+};
